requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-filter
django-extensions
django-cors-headers
drf-yasg
python-dotenv
psycopg2-binary
Pillow
requests
razorpay
locust